Move the code to an utils file.
The most minimal thing to do is to leave the init file totally empty.
