+++ kojak/tests/test_kojak.py
@@ -2,8 +2,8 @@
 import ast
 import unittest
 
-import kojak
-from kojak import Import
+from kojak.utils import (
+    Import, get_classes, get_functions, get_imports, parse_file)
 
 
 sample = '''
@@ -28,13 +28,13 @@ class Bar:
 
 class TestKojak(unittest.TestCase):
     def setUp(self):
-        self.node = kojak.parse_file(sample)
+        self.node = parse_file(sample)
 
     def test_parse_file(self):
         self.assertTrue(isinstance(self.node, ast.Module))
 
     def test_get_imports(self):
-        imports = [el for el in kojak.get_imports(self.node)]
+        imports = [el for el in get_imports(self.node)]
         expected_imports = [
             Import(module=[], name='requests', alias=None),
             Import(module=[], name='abc', alias=None),
@@ -44,12 +44,12 @@ class TestKojak(unittest.TestCase):
         self.assertEqual(imports, expected_imports)
 
     def test_get_classes(self):
-        classes = [el for el in kojak.get_classes(self.node)]
+        classes = [el for el in get_classes(self.node)]
         self.assertEqual(len(classes), 2)
         self.assertEqual(classes[0].name, 'Foo')
         self.assertEqual(classes[1].name, 'Bar')
 
-        methods = [el for el in kojak.get_functions(classes[0].node)]
+        methods = [el for el in get_functions(classes[0].node)]
         self.assertEqual(len(methods), 2)
         self.assertEqual(methods[0].name, 'run')
         self.assertEqual(methods[1].name, 'bar')
